chore: enable newarch support for android and implement module for new and old arch
