+++ packages/repack/android/src/oldarch/java/com/callstack/repack/ChunkManagerModule.kt
@@ -1,0 +1,28 @@
+package com.callstack.repack
+
+import android.os.Handler
+import com.facebook.react.bridge.*
+
+class ScriptManagerModule(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext) {
+    private val remoteLoader: RemoteScriptLoader = RemoteScriptLoader(reactApplicationContext)
+    private val fileSystemLoader: FileSystemScriptLoader = FileSystemScriptLoader(reactApplicationContext)
+
+    private var implementation: ScriptManagerModuleImpl = ScriptManagerModuleImpl()
+
+    override fun getName(): String = ScriptManagerModuleImpl.NAME
+
+    @ReactMethod
+    fun loadScript(scriptId: String, configMap: ReadableMap, promise: Promise) {
+        implementation.loadScript(scriptId, configMap, promise, remoteLoader, fileSystemLoader)
+    }
+
+    @ReactMethod
+    fun prefetchScript(scriptId: String, configMap: ReadableMap, promise: Promise) {
+        implementation.prefetchScript(scriptId, configMap, promise, remoteLoader)
+    }
+
+    @ReactMethod
+    fun invalidateScripts(scriptIds: ReadableArray, promise: Promise) {
+        implementation.invalidateScripts(scriptIds, promise, remoteLoader)
+    }
+}
